+++ relay-server/src/services/processor.rs
@@ -40,6 +40,7 @@ use relay_statsd::metric;
 use relay_system::{Addr, FromMessage, NoResponse, Service};
 use reqwest::header;
 use smallvec::{smallvec, SmallVec};
+use zstd::stream::Encoder as ZstdEncoder;
 
 #[cfg(feature = "processing")]
 use {
@@ -2913,6 +2914,13 @@ fn encode_payload(body: &Bytes, http_encoding: HttpEncoding) -> Result<Bytes, st
             let mut encoder = BrotliEncoder::new(Vec::new(), 0, 5, 22);
             encoder.write_all(body.as_ref())?;
             encoder.into_inner()
+        }
+        HttpEncoding::Zstd => {
+            // Use the fastest compression level, our main objective here is to get the best
+            // compression ratio for least amount of time spent.
+            let mut encoder = ZstdEncoder::new(Vec::new(), 1)?;
+            encoder.write_all(body.as_ref())?;
+            encoder.finish()?
         }
     };