+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Terminate the process when one of the services crashes. ([#4249](https://github.com/getsentry/relay/pull/4249))
 
+**Features**:
+
+- Implement zstd http encoding for Relay to Relay communication. ([#4266](https://github.com/getsentry/relay/pull/4266))
+
 ## 24.11.0
 
 **Breaking Changes**: